requirements.txt
Django==2.*
djangorestframework==3.*
Pillow
https://github.com/danielldt/django-jet/archive/dev.zip
djangorestframework-jwt
django-money
django-ordered-model
django-phonenumber-field
phonenumbers
django-tinymce4-lite
django-cors-headers
django-filter
coreapi
drf-yasg
django-crispy-forms
https://github.com/danielldt/django-easy-audit/archive/master.zip
pandas
numpy
python-dateutil
pytz
django-admin-rangefilter
django-redis
packaging
https://github.com/digi604/django-smart-selects/archive/master.zip
#mysqlclient